refactor(fsd): rename VocFilterBar.tsx → VocListHeader.tsx (W1 fix)
File name now matches the exported symbol VocListHeader.
Update index.ts barrel and legacy stub to point to new path.
