+++ frontend/src/components/voc/VocListHeader.tsx
@@ -1,2 +1,2 @@
 // Moved to features/voc-list-filter/ui/VocFilterBar (Step 4)
-export { VocListHeader } from '@features/voc-list-filter/ui/VocFilterBar';
+export { VocListHeader } from '@features/voc-list-filter/ui/VocListHeader';
